"ads2_m12_a1_1: 0/6 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ ADS-2-Assignments/m12/ADS-2_Exam-2/Solution.java
@@ -13,7 +13,7 @@ public class Solution {
             String[] eValues = sc.nextLine().split(" ");
             ewGraph.addEdge(new Edge(Integer.parseInt(eValues[0]),
             						 Integer.parseInt(eValues[1]),
-                					 Integer.parseInt(eValues[2])));
+                					 Double.parseDouble(eValues[2])));
         }
 		String caseToGo = sc.nextLine();
 		switch (caseToGo) {
